Fixing ssh_client issue, moving process_from_dir back to loadDuns
